Added a base of the futur EntityComponentSystem engine. For now, we can create a world and populate it with systems and entities. Entities can be composed by components.
There is two test classes : FooSystem and FooComponent.
FooSystem process all entities it gets and write on the standard output a message.
FooComponent take a int "foo" in parameter but is not used by any system for now. This is a test that MUST be tried for the next development.